server/requirements.txt
openai
faiss-cpu
numpy
python-dotenv

# API dependencies
flask
flask-cors